# Simplify and optimize Add() and Sub()