```
name                     old time/op  new time/op  delta
_Add/big-6               19.0ns ± 0%  19.0ns ± 0%     ~     (all equal)
_Add/uint256-6           3.73ns ± 0%  2.99ns ± 0%  -19.84%  (p=0.008 n=5+5)
_Sub/big-6               19.0ns ± 0%  19.0ns ± 0%     ~     (all equal)
_Sub/uint256-6           3.52ns ± 0%  2.99ns ± 0%  -15.06%  (p=0.029 n=4+4)
_Sub/uint256_of-6        3.81ns ± 0%  1.82ns ± 0%  -52.08%  (p=0.008 n=5+5)
_AddMod/large/big-6       211ns ± 0%   212ns ± 1%     ~     (p=0.079 n=4+5)
_AddMod/large/uint256-6  45.0ns ± 0%  39.5ns ± 0%  -12.18%  (p=0.000 n=5+4)
_AddMod/small/big-6      64.6ns ± 0%  64.6ns ± 0%     ~     (all equal)
_AddMod/small/uint256-6  21.6ns ± 0%  17.5ns ± 0%  -18.98%  (p=0.008 n=5+5)
```

Requires #27.